refactor(ops): reorganize components into per-page folders and fix replay UX
Extract inline components from 5 page files (4420 → 249 lines total) into
per-page folders under components/ops/ (dashboard, projections, dejaview,
queues, replay-progress) with shared primitives in shared/.

Replay UX fixes:
- Historical run detail page now shows completed/failed/cancelled runs
  instead of dead-end "No active replay found" (adds getReplayRun endpoint,
  error field to ReplayHistoryEntry)
- Phase timeline extracted to shared and added to full progress page
- Auto-discover replaces manual step 3 in wizard
- Single aggregate replay is now independent of the full wizard flow
- Fix tenant count showing 0 when "All tenants" selected
- Remove redundant persistent toast when drawer already opens
