+++ langwatch/src/components/ops/dashboard/index.ts
@@ -1,0 +1,1 @@
+export { OpsDashboardContent } from "./OpsDashboardContent";
