+++ langwatch/src/components/ops/dejaview/index.ts
@@ -1,0 +1,1 @@
+export { DejaViewContent } from "./DejaViewContent";
