+++ langwatch/src/components/ops/queues/index.ts
@@ -1,0 +1,1 @@
+export { QueuesContent } from "./QueuesContent";
