+++ langwatch/src/components/ops/queues/types.ts
@@ -1,0 +1,1 @@
+export type StatusFilter = "all" | "ok" | "blocked" | "stale" | "active";
