+++ langwatch/src/components/ops/shared/ConfirmDialog.tsx
@@ -1,0 +1,45 @@
+import { Button, Text } from "@chakra-ui/react";
+import { Dialog } from "~/components/ui/dialog";
+
+export function ConfirmDialog({
+  open,
+  onClose,
+  onConfirm,
+  title,
+  description,
+  isLoading,
+}: {
+  open: boolean;
+  onClose: () => void;
+  onConfirm: () => void;
+  title: string;
+  description: string;
+  isLoading: boolean;
+}) {
+  return (
+    <Dialog.Root open={open} onOpenChange={(e) => !e.open && onClose()}>
+      <Dialog.Content>
+        <Dialog.Header>
+          <Dialog.Title>{title}</Dialog.Title>
+        </Dialog.Header>
+        <Dialog.Body>
+          <Text textStyle="sm">{description}</Text>
+        </Dialog.Body>
+        <Dialog.Footer>
+          <Button variant="outline" size="sm" onClick={onClose}>
+            Cancel
+          </Button>
+          <Button
+            colorPalette="red"
+            size="sm"
+            onClick={onConfirm}
+            loading={isLoading}
+          >
+            Confirm
+          </Button>
+        </Dialog.Footer>
+        <Dialog.CloseTrigger />
+      </Dialog.Content>
+    </Dialog.Root>
+  );
+}
